requirements.txt
flask
python-dotenv
geminiai
